fix(be): CartItem Id Long 데이터타입 Integer로 변경 (#20)

+++ backend/src/main/java/com/back/domain/cart/repository/CartRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CartRepository extends JpaRepository<Cart, Long> {
+public interface CartRepository extends JpaRepository<Cart, Integer> {
     Optional<Cart> findByUser(User user);
 }
